Renaming `MaybeGas` to `Gas` and exporting it as `svm_types::Gas` (instead of `svm_types::gas::MaybeGas`).
