+++ crates/codec/src/receipt/gas.rs
@@ -1,18 +1,18 @@
 use std::io::Cursor;
 
-use svm_types::gas::MaybeGas;
+use svm_types::Gas;
 
 use crate::{Field, ParseError, ReadExt, WriteExt};
 
-pub fn encode_gas_used(gas: &MaybeGas, w: &mut Vec<u8>) {
+pub fn encode_gas_used(gas: &Gas, w: &mut Vec<u8>) {
     let gas = gas.unwrap_or(0);
 
     w.write_u64_be(gas);
 }
 
-pub fn decode_gas_used(cursor: &mut Cursor<&[u8]>) -> Result<MaybeGas, ParseError> {
+pub fn decode_gas_used(cursor: &mut Cursor<&[u8]>) -> Result<Gas, ParseError> {
     match cursor.read_u64_be() {
-        Ok(gas) => Ok(MaybeGas::with(gas)),
+        Ok(gas) => Ok(Gas::with(gas)),
         Err(..) => Err(ParseError::NotEnoughBytes(Field::GasUsed)),
     }
 }
